+++ wavescapes/__init__.py
@@ -1,5 +1,6 @@
 from .pcv import produce_pitch_class_matrix_from_filename
-from .dft import apply_dft_to_pitch_class_matrix
+from .dft import apply_dft_to_pitch_class_matrix, build_utm_from_one_row, normalize_dft
 from .color import complex_utm_to_ws_utm
 from .draw import Wavescape
-from .general import single_wavescape, all_wavescapes, legend_decomposition
+from .general import single_wavescape, all_wavescapes, single_wavescape_from_pcvs, \
+    all_wavescapes_from_pcvs, legend_decomposition
